Finished Projectiles, added SFX, and sprites - 2 September 2021.

+++ Engine/Input/InputSystem.h
@@ -45,7 +45,7 @@ namespace PhoenixEngine
 		int numKeys;
 
 		Vector2 mousePosition;
-		std::array<Uint8, 3> mouseButtonState;
-		std::array<Uint8, 3> prevMouseButtonState;
+		std::array<Uint32, 3> mouseButtonState;
+		std::array<Uint32, 3> prevMouseButtonState;
 	};
 }
